requirements.txt
# Core app
streamlit>=1.36.0
htbuilder>=0.6.2
python-dotenv>=1.0.1

# LLM + APIs
openai>=1.42.0
langchain>=0.2.11
langchain-openai>=0.1.7
langchain-neo4j>=0.1.6
neo4j>=5.20.0

# Vision / Embeddings
torch>=2.2.0
transformers>=4.42.0
safetensors>=0.4.3
pillow>=10.3.0

# Data utilities
pandas>=2.2.2
numpy>=1.26.4
